Moved sidebar component file close to animated-name component in file tree
